Project milestone completed.
Not sure if there is a simpler method to copy the dollar value to a string with only two digits after the decimal. I found a function that wasn't discussed in Chapter 3, but by Googling found to_string. It worked, but it took awhile to figure out how to display it properly and so I needed to create extra int variables and performed some extra math.